+++ src/Dsll.hs
@@ -27,7 +27,7 @@ import Env
 --import Text.PrettyPrint.HughesPJClass hiding (colon, comma, parens, braces)
 import qualified Data.Map.Strict as M
 import qualified Text.Megaparsec.Char.Lexer as L
-import qualified SubstanceTokenizer         as T
+import qualified Tokenizer         as T
 
 --------------------------------------- DSLL AST -------------------------------
 
@@ -372,7 +372,7 @@ checkDsllStmt e (PdStmt (Pd2Const v)) = let pd = Pred2 $ Prd2 { namepred2 = name
 
 -- TODO Implement typecheckihng for statement notations                                                                         -- notations if needed
 checkDsllStmt e (SnStmt s) =
-   let (from,to) = translatePatterns (fromSn s, toSn s) e
+   let (from,to) = T.translatePatterns (fromSn s, toSn s) e
        newSnr = StmtNotationRule {fromSnr = from, toSnr   = to}
    in e {stmtNotations = newSnr : stmtNotations e}
 
@@ -407,75 +407,6 @@ parseDsll dsllFile dsllIn =
               putStrLn "DSLL Env: \n"
               print env1
               return env1
-
------------------------------- Tokenization ------------------------------------
--- TODO: Maybe should be refactored into a seperate file in the future
--- | Get as an input from and to string notataions and returns refined tokenized
---   versions of them
-
--- | Tokenize the given string using the Substance tokenizer
-tokenize :: String -> [T.Token]
-tokenize = T.alexScanTokens
-
-translatePatterns :: (String,String) -> VarEnv -> ([T.Token],[T.Token])
-translatePatterns (fromStr, toStr) dsllEnv =
-  let from = foldl (refineDSLLToken dsllEnv) [] (tokenize fromStr)
-      patterns = filter notPattern from
-      to = foldl (refinePaternToken patterns) [] (tokenize toStr)
-  in (from,to)
-
-notPattern :: T.Token -> Bool
-notPattern (T.Pattern t) = True
-notPattern token = False
-
-refinePaternToken :: [T.Token] -> [T.Token] -> T.Token -> [T.Token]
-refinePaternToken patterns tokens (T.Var v) =
-   if T.Pattern v `elem` patterns
-    then tokens ++ [T.Pattern v]
-    else tokens ++ [T.Var v]
-
-refinePaternToken patterns tokens t = tokens ++ [t]
-
-
-refineDSLLToken :: VarEnv -> [T.Token] -> T.Token -> [T.Token]
-refineDSLLToken dsllEnv tokens (T.Var v) = if isDeclared v dsllEnv then
-                                          tokens ++ [T.DSLLEntity v]
-                                       else tokens ++ [T.Pattern v]
-refineDSLLToken dsllEnv tokens t = tokens ++ [t]
-
--- |This function identify the pattern vars in the sugared notatation in the
---  StmtNotation in the DSLL
-identifyPatterns :: [T.Token] -> [T.Token] -> [T.Token]
-identifyPatterns tokensSugared tokenDesugared =
-   foldl (identifyPattern tokenDesugared) [] tokensSugared
-
-identifyPattern :: [T.Token] -> [T.Token] -> T.Token -> [T.Token]
-identifyPattern tokenDesugared tokensSugared (T.Var v) =
-   if T.Var v `elem` tokenDesugared then
-     tokensSugared ++ [T.Pattern v]
-  else
-    tokensSugared ++ [T.Var v]
-identifyPattern tokenDesugared tokensSugared token = tokensSugared ++ [token]
-
--- | Retranslate a token list into a program
-reTokenize :: [T.Token] -> String
-reTokenize = foldl translate ""
-
--- | Translation function from a specific token back into String
-translate :: String -> T.Token -> String
-translate prog T.Bind = prog ++ ":="
-translate prog T.NewLine = prog ++ "\n"
-translate prog T.PredEq = prog ++ "<->"
-translate prog T.ExprEq = prog ++ "="
-translate prog T.Comma = prog ++ ","
-translate prog T.Lparen = prog ++ "("
-translate prog T.Rparen = prog ++ ")"
-translate prog T.Space = prog ++ " "
-translate prog (T.Sym c) = prog ++ [c]
-translate prog (T.Var v) = prog ++ v
-translate prog (T.Comment c) = prog ++ c
-translate prog (T.DSLLEntity d) = prog ++ d
-translate prog (T.Pattern p) = prog ++ p
 
 -- --------------------------------------- Test Driver -------------------------
 -- | For testing: first uncomment the module definition to make this module the
