+++ src/paper-eval/linear-algebra.dsl
@@ -17,7 +17,7 @@ predicate Not(p : Prop) : Prop
 
 -- Syntactic Sugar Definition
 
-StmtNotation  "det(v1,v2)" -> "determinant(v1,v2)"
+StmtNotation  "det ( v1, v2 ) " -> "determinant(v1 , v2)   "
 
 /*
 ExprNotation "add($v1 ,$v2)" -> "$v1 + $v2" (at level 2, right associativity)
